Bug 1226166 - Move part of system places.js to shared, r=daleharvey
